# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pizzaroo API",
        "description": "Documentación Swagger para Pizzaroo",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-items": {
            "get": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Obtener todos los ítems de pedidos",
                "operationId": "6a6b106e104d558bfc813be8f4adf2f4",
                "responses": {
                    "200": {
                        "description": "Lista de ítems"
                    }
                }
            },
            "post": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Crear nuevo ítem",
                "operationId": "3390f5f422c701c6e520cc66e9d18037",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "pizza_id",
                                    "quantity",
                                    "price_unit"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "pizza_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "price_unit": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ítem creado exitosamente"
                    }
                }
            }
        },
        "/api/order-items/{id}": {
            "put": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Actualizar ítem de pedido",
                "operationId": "c1778023c9a3ecbc744efb765d68415f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "integer"
                                    },
                                    "pizza_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "price_unit": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ítem actualizado exitosamente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order Items"
                ],
                "summary": "Eliminar ítem de pedido",
                "operationId": "87d21d8e8c19f5d96b514a9aa273dede",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ítem eliminado"
                    }
                }
            }
        },
        "/api/pizzas": {
            "get": {
                "tags": [
                    "Pizzas"
                ],
                "summary": "Listar todas las pizzas",
                "operationId": "e521daabe6a50753134f9e413db64fb1",
                "responses": {
                    "200": {
                        "description": "Lista de pizzas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pizzas"
                ],
                "summary": "Crear una nueva pizza",
                "operationId": "c522059a3e69112cf1bb1752dbdfb9e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "price",
                                    "image_url"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pizza creada exitosamente"
                    }
                }
            }
        },
        "/api/pizzas/{id}": {
            "put": {
                "tags": [
                    "Pizzas"
                ],
                "summary": "Actualizar una pizza existente",
                "operationId": "04d81069150f15278f6d3ac8d272cd5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "image_url": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pizza actualizada correctamente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pizzas"
                ],
                "summary": "Eliminar una pizza",
                "operationId": "dc2258eec4c05ff8699f0e90b5fcb881",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pizza eliminada correctamente"
                    }
                }
            }
        },
...
    },
...
}